Debugged allocation and refactored code for CA - Pratibha, Krisha and Maryam
