+++ src/main/java/com/securian/creditcompass/DataAccess/ClaimRepository.java
@@ -2,11 +2,13 @@ package com.securian.creditcompass.DataAccess;
 
 import com.securian.creditcompass.entities.Claim;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
 import java.util.Calendar;
 import java.util.List;
 
 
+@Repository
 public interface ClaimRepository extends JpaRepository<Claim, Long> {
 
     List<Claim> findClaimsByExaminer(String examiner);
