+++ src/main/java/com/securian/creditcompass/DataAccess/ExaminerRepository.java
@@ -2,8 +2,11 @@ package com.securian.creditcompass.DataAccess;
 
 import com.securian.creditcompass.entities.ClaimsExaminer;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
 import java.util.Optional;
 
+@Repository
 public interface ExaminerRepository extends JpaRepository<ClaimsExaminer, Long> {
     Optional<ClaimsExaminer> findByFirstName(String firstName);
     Optional<ClaimsExaminer> findByUsername(String username);
