--- src/main/java/com/securian/creditcompass/dashboard/DashboardRepository.java
@@ -1,11 +1,0 @@
-package com.securian.creditcompass.dashboard;
-
-import com.securian.creditcompass.entities.Claim;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface DashboardRepository extends JpaRepository<Claim, Integer> {
-    Optional<List<Claim>> findClaimsByExaminer();
-}
